update types for portability
This is a long shot (pun intended)...

We want to be able to build and run on 32-bit systems. Having
`int64t`s creates a build & compatibility issue. However, `long
long` is compiler implementation specific, so although it does
guarantee that the type is "at least 64-bit wide", it seems that
it is not the correct way to go.

We must keep this in mind if something breaks down the road.

+++ src/include/vaccel_id.h
@@ -16,4 +16,4 @@
 
 #include <stdint.h>
 
-typedef int64_t vaccel_id_t;
+typedef long long int vaccel_id_t;
