+++ src/ops/tf.c
@@ -88,7 +88,7 @@ void *vaccel_tf_buffer_get_data(
  *
  * The name will be copied inside the node
  */
-struct vaccel_tf_node *vaccel_tf_node_new(const char *name, int64_t id) {
+struct vaccel_tf_node *vaccel_tf_node_new(const char *name, long long int id) {
 	struct vaccel_tf_node *ret = malloc(sizeof(*ret));
 	if (!ret)
 		return NULL;
@@ -131,7 +131,7 @@ const char *vaccel_tf_node_get_name(struct vaccel_tf_node *node)
  * This will return the id of the node if the node is non-NULL
  * or a negative value otherwise
  */
-int64_t vaccel_tf_node_get_id(struct vaccel_tf_node *node)
+long long int vaccel_tf_node_get_id(struct vaccel_tf_node *node)
 {
 	if (!node)
 		return -VACCEL_EINVAL;
@@ -140,7 +140,7 @@ int64_t vaccel_tf_node_get_id(struct vaccel_tf_node *node)
 }
 
 struct vaccel_tf_tensor *
-vaccel_tf_tensor_new(int nr_dims, int64_t *dims, enum vaccel_tf_data_type type)
+vaccel_tf_tensor_new(int nr_dims, long long int *dims, enum vaccel_tf_data_type type)
 {
 	struct vaccel_tf_tensor *ret;
 
@@ -164,7 +164,7 @@ vaccel_tf_tensor_new(int nr_dims, int64_t *dims, enum vaccel_tf_data_type type)
 
 struct vaccel_tf_tensor *
 vaccel_tf_tensor_allocate(
-	int nr_dims, int64_t *dims,
+	int nr_dims, long long int *dims,
 	enum vaccel_tf_data_type type,
 	size_t total_size
 ) {
